+++ src/main/java/ru/innopolis/stc9/earth_stc9/controllers/users/TeachersController.java
@@ -1,5 +1,6 @@
 package ru.innopolis.stc9.earth_stc9.controllers.users;
 
+import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -7,6 +8,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
+import ru.innopolis.stc9.earth_stc9.exceptions.UserException;
 import ru.innopolis.stc9.earth_stc9.pojo.Role;
 import ru.innopolis.stc9.earth_stc9.pojo.User;
 import ru.innopolis.stc9.earth_stc9.services.IUsersService;
@@ -15,6 +17,7 @@ import java.util.List;
 
 @Controller
 public class TeachersController {
+    private static Logger logger = Logger.getLogger(StudentsController.class);
     private final IUsersService service;
 
     @Autowired
@@ -28,10 +31,15 @@ public class TeachersController {
     }
 
     private String getTeachers(Model model) {
-        model.addAttribute("createRole", Roles.TEACHER_ROLE_ID);
-        List<User> users = service.getUsers(Roles.TEACHER_ROLE_ID);
-        if (users != null) {
-            model.addAttribute("teachers", users);
+        try {
+            model.addAttribute("createRole", Roles.TEACHER_ROLE_ID);
+            List<User> users = service.getUsers(Roles.TEACHER_ROLE_ID);
+            if (users != null) {
+                model.addAttribute("teachers", users);
+            }
+        } catch (Exception e) {
+            logger.warn("Get teachers", e);
+            showMessage(model, e.getMessage());
         }
         return "teachers";
     }
@@ -44,21 +52,39 @@ public class TeachersController {
 
     @RequestMapping(value = "/teacherCreate", method = RequestMethod.POST)
     public String addTeacher(@RequestAttribute String nameTeacher, @RequestAttribute String loginTeacher,
-                             @RequestAttribute String passwordTeacher, @RequestAttribute String roleTeacher,
+                             @RequestAttribute String passwordTeacher, @RequestAttribute Integer roleTeacher,
                              Model model) {
-        User user = new User(nameTeacher, loginTeacher, passwordTeacher, Integer.parseInt(roleTeacher));
-        service.createUser(user);
+        try {
+            if (nameTeacher == null || nameTeacher.isEmpty() || loginTeacher == null || loginTeacher.isEmpty() ||
+                    passwordTeacher == null || passwordTeacher.isEmpty() || roleTeacher == null) {
+                throw new UserException("Ошибка создания преподавателя, переданы нулевые параметры");
+            }
+            User user = new User(nameTeacher, loginTeacher, passwordTeacher, roleTeacher);
+            service.createUser(user);
+        } catch (Exception e) {
+            logger.warn("Create teacher", e);
+            showMessage(model, e.getMessage());
+            addTeacherShowBlock(model);
+        }
         return getTeachers(model);
     }
 
     @RequestMapping(value = "/teacherEdit/{id}", method = RequestMethod.GET)
     public String editTeacherShowBlock(@PathVariable(value = "id") int id, Model model) {
-        User userById = service.getUserById(id);
-        model.addAttribute("user", userById);
-        model.addAttribute("update", "Изменить");
-        List<Role> roles = service.getRoles();
-        if (roles != null) {
-            model.addAttribute("roles", roles);
+        try {
+            if (id <= 0) {
+                throw new UserException("Передан некорректный идентификатор преподавателя");
+            }
+            User userById = service.getUserById(id);
+            model.addAttribute("user", userById);
+            model.addAttribute("update", "Изменить");
+            List<Role> roles = service.getRoles();
+            if (roles != null) {
+                model.addAttribute("roles", roles);
+            }
+        } catch (Exception e) {
+            logger.warn("Edit teacher", e);
+            showMessage(model, e.getMessage());
         }
         return getTeachers(model);
     }
@@ -66,15 +92,38 @@ public class TeachersController {
     @RequestMapping(value = "/teacherEdit/{id}", method = RequestMethod.POST)
     public String editTeacher(@PathVariable(value = "id") int id, @RequestAttribute String nameTeacher,
                               @RequestAttribute String loginTeacher, @RequestAttribute String passwordTeacher,
-                              @RequestAttribute String roleTeacher, Model model) {
-        User user = new User(id, loginTeacher, passwordTeacher, Integer.parseInt(roleTeacher), nameTeacher);
-        service.updateUser(user);
+                              @RequestAttribute Integer roleTeacher, Model model) {
+        try {
+            if (id <= 0 || nameTeacher == null || nameTeacher.isEmpty() || loginTeacher == null ||
+                    loginTeacher.isEmpty() || passwordTeacher == null || passwordTeacher.isEmpty() ||
+                    roleTeacher == null) {
+                throw new UserException("Ошибка изменения преподавателя, переданы нулевые параметры");
+            }
+            User user = new User(id, loginTeacher, passwordTeacher, roleTeacher, nameTeacher);
+            service.updateUser(user);
+        } catch (Exception e) {
+            logger.warn("Edit teacher", e);
+            showMessage(model, e.getMessage());
+            editTeacherShowBlock(id, model);
+        }
         return getTeachers(model);
     }
 
     @RequestMapping(value = "/teacherDelete/{id}", method = RequestMethod.POST)
     public String deleteTeacher(@PathVariable(value = "id") int id, Model model) {
-        service.deleteUserById(id);
+        try {
+            if (id <= 0) {
+                throw new UserException("Передан некорректный идентификатор");
+            }
+            service.deleteUserById(id);
+        } catch (Exception e) {
+            logger.warn("Delete teacher", e);
+            showMessage(model, e.getMessage());
+        }
         return getTeachers(model);
+    }
+
+    private void showMessage(Model model, String message) {
+        model.addAttribute("message", message);
     }
 }
