+++ src/main/java/ru/innopolis/stc9/earth_stc9/controllers/users/UsersController.java
@@ -1,5 +1,6 @@
 package ru.innopolis.stc9.earth_stc9.controllers.users;
 
+import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -7,6 +8,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
+import ru.innopolis.stc9.earth_stc9.exceptions.UserException;
 import ru.innopolis.stc9.earth_stc9.pojo.Role;
 import ru.innopolis.stc9.earth_stc9.pojo.User;
 import ru.innopolis.stc9.earth_stc9.services.IUsersService;
@@ -15,6 +17,7 @@ import java.util.List;
 
 @Controller
 public class UsersController {
+    private static Logger logger = Logger.getLogger(UsersController.class);
     private final IUsersService service;
 
     @Autowired
@@ -28,40 +31,69 @@ public class UsersController {
     }
 
     private String getUsers(Model model) {
-        List<User> users = service.getUsers();
-        if (users != null) {
-            model.addAttribute("users", users);
+        try {
+            List<User> users = service.getUsers();
+            if (users != null) {
+                model.addAttribute("users", users);
+            }
+        } catch (Exception e) {
+            logger.warn("Get users", e);
+            showMessage(model, e.getMessage());
         }
         return "users";
     }
 
     @RequestMapping(value = "/userCreate", method = RequestMethod.GET)
     public String addUserShowBlock(Model model) {
-        model.addAttribute("create", "Создать");
-        List<Role> roles = service.getRoles();
-        if (roles != null) {
-            model.addAttribute("roles", roles);
+        try {
+            model.addAttribute("create", "Создать");
+            List<Role> roles = service.getRoles();
+            if (roles != null) {
+                model.addAttribute("roles", roles);
+            }
+
+        } catch (Exception e) {
+            logger.warn("Create user", e);
+            showMessage(model, e.getMessage());
         }
         return getUsers(model);
     }
 
     @RequestMapping(value = "/userCreate", method = RequestMethod.POST)
     public String addUser(@RequestAttribute String nameUser, @RequestAttribute String loginUser,
-                          @RequestAttribute String passwordUser, @RequestAttribute String role,
+                          @RequestAttribute String passwordUser, @RequestAttribute Integer role,
                           Model model) {
-        User user = new User(nameUser, loginUser, passwordUser, Integer.parseInt(role));
-        service.createUser(user);
+        try {
+            if (nameUser == null || nameUser.isEmpty() || loginUser == null || loginUser.isEmpty() ||
+                    passwordUser == null || passwordUser.isEmpty() || role == null) {
+                throw new UserException("Ошибка создания пользователя, переданы нулевые параметры пользователя");
+            }
+            User user = new User(nameUser, loginUser, passwordUser, role);
+            service.createUser(user);
+        } catch (Exception e) {
+            logger.warn("Create user", e);
+            showMessage(model, e.getMessage());
+            return addUserShowBlock(model);
+        }
         return getUsers(model);
     }
 
     @RequestMapping(value = "/userEdit/{id}", method = RequestMethod.GET)
     public String editUserShowBlock(@PathVariable(value = "id") int id, Model model) {
-        User userById = service.getUserById(id);
-        model.addAttribute("user", userById);
-        model.addAttribute("update", "Изменить");
-        List<Role> roles = service.getRoles();
-        if (roles != null) {
-            model.addAttribute("roles", roles);
+        try {
+            if (id <= 0) {
+                throw new UserException("Передан некорректный идентификатор");
+            }
+            User userById = service.getUserById(id);
+            model.addAttribute("user", userById);
+            model.addAttribute("update", "Изменить");
+            List<Role> roles = service.getRoles();
+            if (roles != null) {
+                model.addAttribute("roles", roles);
+            }
+        } catch (Exception e) {
+            logger.warn("Edit user", e);
+            showMessage(model, e.getMessage());
         }
         return getUsers(model);
     }
@@ -69,15 +101,37 @@ public class UsersController {
     @RequestMapping(value = "/userEdit/{id}", method = RequestMethod.POST)
     public String editUser(@PathVariable(value = "id") int id, @RequestAttribute String nameUser,
                            @RequestAttribute String loginUser, @RequestAttribute String passwordUser,
-                           @RequestAttribute String role, Model model) {
-        User user = new User(id, loginUser, passwordUser, Integer.parseInt(role), nameUser);
-        service.updateUser(user);
+                           @RequestAttribute Integer role, Model model) {
+        try {
+            if (id <= 0 || nameUser == null || nameUser.isEmpty() || loginUser == null || loginUser.isEmpty() ||
+                    passwordUser == null || passwordUser.isEmpty() || role == null) {
+                throw new UserException("Ошибка изменения пользователя, переданы нулевые параметры пользователя");
+            }
+            User user = new User(id, loginUser, passwordUser, role, nameUser);
+            service.updateUser(user);
+        } catch (Exception e) {
+            logger.warn("Edit user", e);
+            showMessage(model, e.getMessage());
+            editUserShowBlock(id, model);
+        }
         return getUsers(model);
     }
 
     @RequestMapping(value = "/userDelete/{id}", method = RequestMethod.POST)
     public String deleteUser(@PathVariable(value = "id") int id, Model model) {
-        service.deleteUserById(id);
+        try {
+            if (id <= 0) {
+                throw new UserException("Ошибка удаления пользователя, передан некорректный идентификатор");
+            }
+            service.deleteUserById(id);
+        } catch (Exception e) {
+            logger.warn("Delete user", e);
+            showMessage(model, e.getMessage());
+        }
         return getUsers(model);
+    }
+
+    private void showMessage(Model model, String message) {
+        model.addAttribute("message", message);
     }
 }
